This is a Socket.IO/WebRTC Chat without much else to finish coding. I have a newer version I am working on.
So first if you see anything you can fix it or finish it yourself. I coded most of what I saw floating around in people's minds about wanting a chat with a backend that operates as a chatbot does. Well, here you guys go and hope to see some people finish up my work. run project however but I host the entire website with this method. You may have to uninstall apache or an alt method if you do not change the ports.